* Put the code in its own namespace * Added doxygen comments to all places that doxygen complained they   were missing * Renamed most of the variables in the code.
At this point basic functionality has been verified. Obviously we
could keep working on naming conventions and comments for a long
long time more, but let's just stop here

Immediate attention to be given to mosquitto on connect/reconnect
callbacks - subsxriptions are lost in case of disconnection with
the broker.